+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to `ionesculiviucristian/laravel-html-form` will be documented in this file
 
+## 1.1.1 - 2019-04-16
+- added get/post/put/delete methods to Form
+- added jsClass() to Element
+
 ## 1.1.0 - 2019-04-16
 
 - refactored code to allow better Element extensions